chore: remove duplicate code
