+++ src/app/collective-rewards/rewards/hooks/useGetBuilderRewardPercentage.ts
@@ -2,14 +2,9 @@ import { BackersManagerAbi } from '@/lib/abis/v2/BackersManagerAbi'
 import { AVERAGE_BLOCKTIME } from '@/lib/constants'
 import { BackersManagerAddress } from '@/lib/contracts'
 import { useEffect, useState } from 'react'
-import { Address, parseEther } from 'viem'
+import { Address } from 'viem'
 import { useReadContract } from 'wagmi'
-
-interface BuilderRewardPercentage {
-  current: number
-  next: number
-  cooldownEndTime: bigint
-}
+import { BuilderRewardPercentage, getPercentageData } from '../utils/getPercentageData'
 
 export const useGetBuilderRewardPercentage = (builder: Address) => {
   const [rewardPercentageData, setRewardPercentageData] = useState<BuilderRewardPercentage>()
@@ -28,17 +23,7 @@ export const useGetBuilderRewardPercentage = (builder: Address) => {
 
     const [previous, next, cooldownEndTime] = data
 
-    const currentTimestamp = Math.floor(Date.now() / 1000)
-    const previousPercentage = Number((previous * 100n) / parseEther('1'))
-    const nextPercentage = Number((next * 100n) / parseEther('1'))
-    let currentPercentage = currentTimestamp < cooldownEndTime ? previousPercentage : nextPercentage
-    currentPercentage = Math.round(currentPercentage * 100) / 100
-
-    const percentageData: BuilderRewardPercentage = {
-      current: currentPercentage,
-      next: nextPercentage,
-      cooldownEndTime,
-    }
+    const percentageData = getPercentageData(previous, next, cooldownEndTime)
 
     setRewardPercentageData(percentageData)
   }, [data])
