+++ src/app/collective-rewards/rewards/hooks/useGetBuildersRewardPercentage.ts
@@ -2,14 +2,10 @@ import { BackersManagerAbi } from '@/lib/abis/v2/BackersManagerAbi'
 import { AVERAGE_BLOCKTIME } from '@/lib/constants'
 import { BackersManagerAddress } from '@/lib/contracts'
 import { useMemo } from 'react'
-import { Address, parseEther } from 'viem'
+import { Address } from 'viem'
 import { useReadContracts } from 'wagmi'
-
-export interface BuilderRewardPercentage {
-  current: number
-  next: number
-  cooldownEndTime: bigint
-}
+import { BuilderRewardPercentage, getPercentageData } from '../utils/getPercentageData'
+import { toPercentage } from '../utils/toPercentage'
 
 export const useGetBuildersRewardPercentage = (builders: Address[]) => {
   const rewardPercentageCalls = useMemo(
@@ -39,20 +35,7 @@ export const useGetBuildersRewardPercentage = (builders: Address[]) => {
     () =>
       rewardSharesResult
         ?.map(share => share.result as [bigint, bigint, bigint])
-        .map(([previous, next, cooldownEndTime]) => {
-          const currentTimestamp = Math.floor(Date.now() / 1000)
-          const previousPercentage = Number((previous * 100n) / parseEther('1'))
-          const nextPercentage = Number((next * 100n) / parseEther('1'))
-          let currentPercentage = currentTimestamp < cooldownEndTime ? previousPercentage : nextPercentage
-          currentPercentage = Math.round(currentPercentage * 100) / 100
-
-          const percentageData: BuilderRewardPercentage = {
-            current: currentPercentage,
-            next: nextPercentage,
-            cooldownEndTime,
-          }
-          return percentageData
-        }),
+        .map(([previous, next, cooldownEndTime]) => getPercentageData(previous, next, cooldownEndTime)),
     [rewardSharesResult],
   )
