+++ src/app/collective-rewards/rewards/utils/toPercentage.ts
@@ -1,0 +1,3 @@
+import { parseEther } from 'viem'
+
+export const toPercentage = (value: bigint) => Number((value * 100n) / parseEther('1'))
